move tests with their dependencies to their own meta-module with go.mod
