+++ pkg/chanstreaming/go.mod
@@ -1,11 +1,3 @@
 module github.com/diemenator/go-chanstreaming/pkg/chanstreaming
 
 go 1.24.1
-
-require github.com/stretchr/testify v1.10.0
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
